Make EditText.highlightText faster by adding spans directly
The previous code called setText() after adding each span, which is
very slow when there is a lot of text. It is enough to call setText()
once at the end if any matches were found.

But we can do even better: we can add the highlighting spans directly
to the existing Editable of the EditText, and skip setText() entirely,
thus avoiding any copying of the text.

The documentation of EditText.getText() says "The content of the
return value should not be modified. If you want a modifiable one, you
should make your own copy first." (inherited from TextView), but
adding spans to the return value appears to work fine.