Update Kotlin to 1.9.23
This fixes the build after the recent update of the Composer Compiler:

e: This version (1.5.11) of the Compose Compiler requires Kotlin version 1.9.23 but you appear to be using Kotlin version 1.9.22 which is not known to be compatible.  Please consult the Compose-Kotlin compatibility map located at https://developer.android.com/jetpack/androidx/releases/compose-kotlin to choose a compatible version pair (or `suppressKotlinVersionCompatibilityCheck` but don't say I didn't warn you!).